+++ api/src/user/schema.rs
@@ -6,10 +6,10 @@ pub struct CreateUserSchema {
     pub password: String,
     pub email: String,
     pub fullname: String,
-    pub role: i8,
+    pub role: i16,
     pub avatar: String,
     pub notes: String,
-    pub active: i8,
+    pub active: i16,
 }
 #[derive(Serialize, Deserialize, Debug)]
 pub struct UpdateUserSchema {
@@ -17,8 +17,8 @@ pub struct UpdateUserSchema {
     pub password: Option<String>,
     pub email: Option<String>,
     pub fullname: Option<String>,
-    pub role: Option<i8>,
+    pub role: Option<i16>,
     pub avatar: Option<String>,
     pub notes: Option<String>,
-    pub active: Option<i8>,
+    pub active: Option<i16>,
 }
